+++ tests/server/blueprints/institutes/test_institute_views.py
@@ -420,6 +420,7 @@ def test_institute_settings(app, user_obj, institute_obj):
                 "HP:0001250 , Seizures ( EP )",
             ],
             "gene_panels": [test_panel["panel_name"]],
+            "alamut_key": "test_alamut_key",
         }
 
         # via POST request
@@ -441,6 +442,7 @@ def test_institute_settings(app, user_obj, institute_obj):
         assert updated_institute["gene_panels"] == {
             test_panel["panel_name"]: test_panel["display_name"]
         }
+        assert updated_institute["alamut_key"] == form_data["alamut_key"]
 
 
 def test_cases(app, institute_obj):