+++ CHANGELOG.md
@@ -19,6 +19,8 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Added GMS and Lund university hospital logos to login page
 - Made display of Swedac logo configurable
 - Support for displaying custom images in case view
+- Individual-specific HPO terms
+- Optional alamut_key in institute settings for Alamut Plus software
 ### Fixed
 - Updated IGV to v2.8.5 to solve missing gene labels on some zoom levels
 - Demo cancer case config file to load somatic SNVs and SVs only.